project/task-manager/requirements.txt
asgiref==3.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Deprecated==1.2.13
Django==3.2
django-oauth-toolkit==1.7.0
djangorestframework==3.13.1
idna==3.3
jwcrypto==1.2
kafka-python==2.0.2
oauthlib==3.2.0
psycopg2==2.9.3
pycparser==2.21
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
wrapt==1.14.1
